Move bundler.go to the docs package
